+++ output.json
@@ -1,0 +1,4 @@
+{
+   "filename": "1_prac.txt",
+   "contents": "에르나는 변호사와의 편지로 인해 재산 상속을 받지 못할 것을 알게 되고, 이로 인해 절망과 분노에 휩싸이면서도 토마스 바덴의 제안에 대해 선택과 고민을 하게 된다."
+}
