+++ uploads/1_prac.txt
@@ -1,0 +1,197 @@
+1. 봄날의 독버섯
+
+에르나 하르디는 착한 아이였다.
+
+착한 숙녀로 자랐으니 이제 곧 착한 아내가 될 차례였다.
+
+장황한 편지의 내용을 요약하자면 그랬다. 그러니 특별히 너에게 내 착한 아내가 될 기회를 주겠다는, 기대한 것과는 달라도 한참 다른 답장이었다.
+
+“말도 안 돼.”
+
+한참을 노려보던 편지를 내려놓는 에르나의 동작은 무심코 내뱉은 혼잣말처럼 단호했다.
+
+“이건 정말 말도 안 돼!”
+
+다시 한번 생각해 보아도 결론은 마찬가지. 벌떡 자리에서 일어선 에르나는 종종걸음을 쳐 창가로 다가섰다. 쏟아지는 봄볕이 눈부셔 더욱 울적해지는 오후였다.
+
+삐거덕거리는 들창을 연 에르나는 창틀에 올라앉아 두 무릎을 감싸 안았다. 지대가 높은 곳에 자리한 바덴 가의 저택에서는 마을의 정경이 한눈에 내려다보였다. 사과꽃이 물결치는 과수원과 개울, 노란 앵초로 뒤덮인 완만한 경사의 언덕을 느리게 배회하던 에르나의 시선은 정원 한편에 놓여 있는 주인 잃은 의자 위에서 멈추었다.
+
+세상은 한 인간의 불행 따위에는 아무 관심이 없다.
+
+그 당연한 사실이 에르나는 새삼 쓰라렸다. 사랑하는 가족을 잃고, 삶의 터전에서 내쫓길 위기에 처해도 완연한 봄기운이 가득한 세상은 무정하게 아름답다. 만약 이 바보 같은 푸념을 듣는다면 할아버지는 그저 웃으실 테지. 그러니 얼마나 다행이냐. 약간의 냉소가 섞인 태평한 말을 덧붙이며.
+
+“아가씨! 에르나 아가씨!”
+
+상념에 잠긴 에르나를 일깨운 건 복도 저편에서부터 들려오기 시작한 가정부 그레베 부인의 목소리였다. 벌써 점심 식사를 할 시간이 된 모양이었다.
+
+“네에! 가요! 곧 내려갈게요!”
+
+목청 높여 대답한 에르나는 서둘러 창틀에서 내려섰다. 말도 안 되는 편지는 눈에 띄지 않도록 서랍 속에 숨겨 두고, 흐트러진 옷매무새도 가다듬었다.
+
+괜찮아.
+
+뛰다시피 해 1층 식당을 향해 가는 내내 에르나는 주문을 외우듯 중얼거렸다.
+
+괜찮아, 다 괜찮을 거야.
+
+* * *
+
+“에르나, 혹시 변호사는 만나 보았니?”
+
+오늘의 날씨와 새로 만든 조각보 같은 잡담을 열의 없이 늘어놓던 바덴 남작부인은 식사를 끝마칠 때가 되어서야 넌지시 본론을 꺼냈다. 침착한 태도를 유지하려 애쓰는 모습이었지만 눈빛에서는 숨길 수 없는 초조함이 묻어났다.
+
+“아니요, 할머니. 아직이요.”
+
+에르나는 단호한 어조로 서둘러 대답했다.
+
+“이번 주가 가기 전에는 꼭 만나 보도록 할게요.”
+
+목과 허리를 꼿꼿이 세운 에르나 위로 창문을 지나온 햇빛이 내려앉았다. 심장의 박동 소리가 어찌나 크게 들려오는지. 입술이 마르고 손끝이 다 곱아들었다. 다행히 바덴 남작부인은 더 이상의 것을 캐묻지 않고 고개를 끄덕여 주었다.
+
+“그래. 그래야지. 부디 방법을 찾을 수 있어야 할 텐데 말이다.”
+
+나직한 한숨 소리가 지나치게 고요한 식당의 공기 속으로 스며들었다.
+
+에르나는 무릎 위에 놓인 두 손을 응시하던 눈길을 들어 물끄러미 할머니를 바라보았다. 한 달 사이에 부쩍 늙고 쇠약해진 모습이었다. 하루아침에 남편을 잃고, 얼마 남지 않은 재산마저 몽땅 남과 다름없는 친척에게 넘겨주게 생겼으니 그럴 만도 했다. 그러니 어떻게 진실을 말할 수 있겠는가.
+
+에르나는 이미 충분히 반듯한 자세를 고쳐 앉으며 마른침을 삼켰다. 힘을 주어 꼭 다문 입술에는 절대 진실을 발설하지 않겠다는 고집이 가득 담겨 있었다.
+
+실은 진작 변호사를 만났다. 돌아온 대답은 에르나가 이미 알고 있는 것과 조금도 다르지 않았다.
+
+아들이 없는 바덴 남작의 재산은 조카에게 상속된다.
+
+그런 말도 안 되는 법이 존재한다는 건 에르나도 진작부터 알고 있었다. 속상하고 억울한 일이지만, 법을 바꿀 길이 없다면 대비책을 마련해야 했다. 에르나가 일을 조금씩 늘려 여윳돈을 모으기 시작한 건 그런 결심을 한 무렵부터였다. 언젠가 남의 손에 넘어갈 이 집을 정당하게 다시 사들일 수 있도록. 하지만 그 ‘언젠가’는 지나치게 빨리 찾아와 버렸고, 모은 돈은 터무니없이 적었다.
+
+'안타깝게 되었지만 상속법이란 게 본래 그렇습니다, 하르디 양.’
+
+다른 방법이 없겠느냐고 애걸하듯 묻는 에르나에게 변호사는 무성의한 대꾸로만 일관했다.
+
+‘지금으로서는 바덴 씨에게 사정을 설명하고 자비를 구하는 것이 최선일 것 같군요.’
+
+그 말을 끝으로 그는 다시 파이프를 입에 물었다. 무례하기 짝이 없는 태도였지만 에르나는 꾹 참아 냈다. 상담료를 제대로 지불하기도 힘든 처지에 있는 고객을 환대하는 변호사는 그리 많지 않을 테니까.
+
+그날 오후, 에르나는 토마스 바덴에게 보내는 편지를 썼다. 아무리 생각해 보아도 변호사가 말한 최선 이상을 찾아내기 힘든 탓이었다. 그리고 오늘, 토마스 바덴이 보낸 답장은 실낱같던 희망을 까마득한 절망과 분노로 바꾸어 놓았다.
+
+“다 잘될 거예요, 할머니. 너무 걱정하지 마세요.”
+
+할머니를 안심시키기 위한 거짓말과 미소를 건넨 에르나는 그만 식탁에서 일어나 앞치마를 둘렀다. 느릿느릿 다가온 그레베 부인을 도와 식탁을 정리하는 에르나의 손놀림은 무척이나 능숙했다.
+
+괜찮지 않다.
+
+식기를 북북 문질러 닦으며 에르나는 더 이상 회피할 수 없는 진실을 받아들였다.
+
+몰락한 귀족인 바덴 남작가의 재산은 이제 이 시골 저택 한 채가 전부라 해도 과언이 아니었다. 하지만 이 집은 곧 적법한 상속자라는 토마스 바덴의 차지가 될 터였다. 그리고 그는 일말의 고민도 없이 이 땅을 팔아 치우겠지.
+
+에르나는 숨을 깊이 들이쉬어 치밀어 오르는 울분을 억눌렀다. 거친 손놀림을 따라 튀어 오른 비누 거품이 걷어 올린 소매 끝과 앞치마에 자잘한 얼룩을 남겼다.
+
+토마스 바덴은 에르나의 마음을 충분히 이해한다고 말했다. 하지만 자신에게도 나름의 사정이 있으므로 바덴 남작부인이 세상을 떠날 때까지 저택의 처분을 미루어 줄 수는 없다고.
+
+차라리 그처럼 단호한 거절의 뜻만 밝혔다면 이런 기분이 들지는 않았을 텐데.
+
+설거지를 마친 에르나는 아무렇게나 벗은 앞치마를 말아 쥔 채 뒤뜰로 향했다. 아름드리 물푸레나무 아래에 놓인 할아버지의 의자에 앉자 눈시울이 뜨겁게 달아올랐다.
+
+그 말도 안 되는 답장에는 토마스 바덴이 제시한 한 가지 타협안이 담겨 있었다. 만약 에르나 하르디가 자신의 아내가 된다면 특별히 관용을 베풀어 줄 의사가 있다고 말이다.
+
+아름다운 봄의 정경이 담긴 시야가 흐릿해지기 시작했지만 에르나는 두 눈을 부릅떠 눈물을 참아 냈다. 고작 그런 사람 때문에 울고 싶지는 않았다. 궁지에 내몰린 친척을 이런 식으로 대하는, 아버지뻘은 족히 되면서도 나잇값을 하지 못하는 그런…….
+
+“아버지.”
+
+에르나는 저도 모르게 중얼거렸다. 오랜 세월 잊고 살아왔지만 그래도 분명 이 세상에 존재하기는 하는 이름을.
+
+그래, 아버지!
+
+두 눈을 가늘게 찌푸리며 에르나는 의자에서 벌떡 일어섰다. 떨어진 앞치마의 끈이 팔랑, 봄바람의 결을 따라 나부꼈다.
+
+* * *
+
+비에른을 깨운 건 닫힌 창문과 두꺼운 커튼도 막지 못한 바깥의 소음이었다. 대공저 옆을 흐르는 강에서 시작된 떠들썩한 기합과 함성이 어둑한 침실까지 비집고 들어왔다.
+
+베개와 쿠션 속에 머리를 파묻은 채 다시 잠을 청해 보았지만 비에른은 결국 굴복했다.
+
+“기운 넘치는 미친 새끼들.”
+
+비에른은 한숨과 함께 욕지거리를 내뱉으며 침대를 빠져나왔다. 서편의 창문을 가린 커튼을 열자 강 저편에서 한창 조정 연습을 하는 무리가 보였다.
+
+매년 여름, 이 도시를 가로질러 바다로 흘러드는 아비트 강에서는 귀족들의 조정 경기가 열린다. 파티와 가십만으로 버텨 내기에는 여름이 지나치게 길어 지겨우니 뭐라도 해 보려는 노력은 가상하지만, 문제는 강이 대공저와 가깝다는 데 있었다. 본격적인 연습이 시작되는 봄부터 경기가 끝나는 여름까지는 이 끔찍한 소음에서 벗어나기 힘들었다.
+
+창틀에 비스듬히 몸을 기대선 비에른은 좁아터진 보트에 징그럽게 붙어 앉아 도무지 이해할 수 없는 열정을 불태우는 사내새끼들을 물끄러미 바라보았다.
+
+넘쳐 나는 기운이 주체가 안 되면 차라리 섹스나 하지, 미친놈들.
+
+저 쓸데없이 땀 빼는 덧없는 짓거리보다는 그편이 훨씬 이로운 취미겠다. 최악의 경우가 생긴다 해도 아이 하나쯤은 남길 테니 왕국의 인구 증가로 인한 국력 상승에 사소한 기여라도 하겠지. 물론 개인사에 있어서는 골머리 썩는 일이겠지만 조절 못 하고 싸는 병신들의 비극은 그의 알 바 아닌 일이었다.
+
+테이블에 놓여 있는 미지근한 물을 한 모금 마신 비에른은 헝클어진 머리를 쓸어 넘기며 돌아섰다. 가운을 대충 걸쳐 입고 종을 울리자 곧 집사 그렉이 들었다.
+
+“죄송합니다, 왕자님. 궁의 사유지 사용 요청에는 응하지 않았지만 그 인근까지 이용하는 건 이미 슈베린 시청의 허가가 나 막을 길이 없습니다.”
+
+대공저에서는 한밤중이나 다름없는 정오에 종이 울린 이유를 잘 아는 그렉은 주인이 묻기 전에 서둘러 보고했다.
+
+“올해는 참가하는 팀이 늘어 조금 더 소란해질 모양입니다.”
+
+그가 덧붙인 비극적인 소식에 비에른은 실소를 터트렸다.
+
+“어차피 우승은 레오니트 드나이스터가 하고 다채로운 얼간이들이 들러리 서는 짓을 참 열성적이게들 하시네.”
+
+“침실을 옮기시겠습니까?”
+
+“아니. 괜찮아요.”
+
+“그러면 식사 준비를 하겠습니다.”
+
+“발코니로. 과일 정도만.”
+
+건조한 명령을 남긴 비에른은 욕실로 들어섰다. 뜨거운 물로 오랫동안 샤워를 하고 나오자 침실의 발코니에 차려진 식탁이 그를 기다리고 있었다.
+
+비에른은 차가운 위스키소다를 마시며 발아래로 펼쳐진 풍경을 응시했다. 슈베린 궁의 명물이라 불리는 대분수의 물줄기가 시원스레 뿜어져 나오고 있었다. 분수대를 장식하고 있는 금빛 동상들과 부서지는 포말이 화창한 봄볕 속에서 빛났다.
+
+비에른의 시선은 대공저와 정원을 잇는 계단의 사면을 따라 이어진 분수를 지나 그 물줄기가 흘러가는 수로에 닿았다. 길게 뻗은 수로의 끝과 닿아 있는 아비트 강에서는 여전히 우렁찬 기합 소리가 들려오고 있었다.
+
+“왕자님, 왕세자 전하께서 드셨습니다.”
+
+다급하게 다가온 그렉이 전언한 건 비에른이 막 얼음만 남은 잔을 테이블 위에 내려놓았을 무렵이었다.
+
+냅킨으로 손가락에 남은 물기를 닦아 낸 비에른은 사과 한 알을 집어 들며 건성으로 고개를 끄덕였다. 집사가 물러간 지 얼마 지나지 않아 침실에 든 레오니트는 성큼성큼 다가와 그의 맞은편에 자리했다. 조정 연습을 하다 달려온 것이 분명한 차림새였다.
+
+“왕세자 전하를 뵙습니다.”
+
+비에른은 다리를 꼬아 앉은 오만한 자세와는 딴판으로 우아한 인사로 동생을 맞이했다. 고개를 까딱이는 리듬을 따라 백금빛의 머리칼 끝에 맺혀 있던 물방울이 떨어져 내렸다.
+
+기가 막힌다는 듯이 쳐다보는 레오니트를 곁에 두고도 비에른은 아무렇지 않게 정원의 대분수를 내려다보았다. 한 입 크게 베어 문 아삭한 과육은 바람을 타고 맴도는 꽃향기처럼 달콤했다.
+
+“용건을 말해.”
+
+왕세자 몫의 차를 준비한 하인들이 물러가고 나자 비에른은 가늘게 뜬 눈으로 레오니트를 마주했다.
+
+대공의 궁전이 자리하고 있는 슈베린은 레첸의 귀족들이 여름을 보내기 위해 찾는 휴양 도시이기도 했다. 아직은 이른 때였지만, 조정에 미친 왕세자는 일찌감치 슈베린 궁전으로 옮겨 와 형의 일상을 헤집어 놓는 데 큰 일조를 하고 있었다.
+
+가벼운 한숨을 흘린 레오니트는 가져온 신문을 테이블 위에 툭 내려놓는 것으로 대답을 대신했다. 사교계의 가십을 전문적으로 다루는 유명 타블로이드지의 첫 면에는 단골손님인 비에른 드나이스터의 사진이 큼지막하게 실려 있었다.
+
+왕실의 독버섯, 이대로 괜찮은가.
+
+우스꽝스러운 헤드라인을 훑은 비에른의 미간이 슬쩍 찌푸려졌다.
+
+“독버섯?”
+
+“몰랐어? 대공의 새 별명인 모양이던데?”
+
+독버섯.
+
+느릿하게 되뇌어 본 비에른은 실소하며 신문을 내려놓았다. 그래도 꽤 잘 나온 사진을 써 주었다는 점 하나는 칭찬해 줄 만한 기사였다.
+
+“글래디스가 레첸으로 돌아온대.”
+
+가만히 비에른의 옆얼굴을 바라보던 레오니트가 조심스럽게 말문을 열었다. 글래디스. 그 이름이 일으킨 파문이 비에른의 입술 끝에 걸린 웃음을 지웠다.
+
+제대로 읽어 본 타블로이드지의 기사에는 올여름을 레첸에서 보내기로 한 라르스의 공주 글래디스 하트퍼드에 관한 소식도 제법 상세하게 담겨 있었다.
+
+한때는 온 레첸의 사랑을 받았던 아름다운 왕세자비. 하지만 남편에게 배신당한 채 버려지고, 아이마저 잃은 그 비운의 여인의 귀환이라. 과연 호사가들이 열광할 만한 가십거리였다. 한때는 왕세자였으나 지금은 독버섯으로 전락한 그녀의 전남편까지 더해지면 더더욱 그럴싸한 그림이 만들어질 터였다.
+
+“어쩔 거야, 비에른?”
+
+“글쎄.”
+
+심각한 레오니트를 우습게 만들 만큼 비에른의 태도는 심상했다.
+
+한 입 더 베어 문 사과를 그만 내려놓은 비에른은 의자 깊이 등을 기대앉은 채 손가락을 타고 흐른 과즙을 닦아 냈다. 별다른 감정이 담기지 않은 눈길은 그저 담담할 뿐이었다.
+
+바야흐로 봄. 독버섯이 쑥쑥 자라나기 좋은 계절이었다.
